search-api: Move most of `Language` to `search` lib
